requirements.txt
Django==2.1.5
Pillow==5.4.1
django-registration==3.0
djangorestframework==3.9.1
